refactor: syncing our changes across repo and handler
