+++ server/internal/repository/business_test.go
@@ -2,24 +2,11 @@ package repository_test
 
 import (
 	"context"
-	"database/sql"
-	"os"
 	"testing"
 
 	"github.com/kiefbc/sooke_app/server/internal/repository"
-	"github.com/kiefbc/sooke_app/server/internal/testdb"
 	"github.com/kiefbc/sooke_app/server/internal/testdb/seeds"
 )
-
-var testDB *sql.DB
-
-func TestMain(m *testing.M) {
-	testDB = testdb.Open()
-	if testDB == nil {
-		os.Exit(0)
-	}
-	os.Exit(m.Run())
-}
 
 func TestGetBusinessBySlug(t *testing.T) {
 	tests := []struct {
